+++ src/controllers/mod.rs
@@ -18,7 +18,7 @@
 use crate::app::core::{CurrentView, MainState};
 use crate::app::model::{
     ConsoleModel, FileInputModel, HomeModel, IndiceModel, InfoAggiuntiveModel, OutputModel,
-    SecondModel, SubModel,
+    HelpModel, SubModel,
 };
 use crate::domain::index::Indice;
 use crate::state::GLOBAL_STATE;
@@ -26,6 +26,8 @@ use raylib::RaylibHandle;
 
 pub(crate) mod home;
 use home::HomeController;
+pub(crate) mod help;
+use help::HelpController;
 pub(crate) mod indice;
 use indice::IndiceController;
 pub(crate) mod file_input;
@@ -39,7 +41,7 @@ use console::ConsoleController;
 
 pub(crate) struct Controllers {
     pub(crate) home_controller: HomeController,
-    pub(crate) second_controller: SecondController,
+    pub(crate) help_controller: HelpController,
     pub(crate) indice_controller: IndiceController,
     pub(crate) fileinput_controller: FileInputController,
     pub(crate) infoaggiuntive_controller: InfoAggiuntiveController,
@@ -51,7 +53,7 @@ impl Controllers {
     pub(crate) fn new() -> Self {
         Self {
             home_controller: HomeController::new(),
-            second_controller: SecondController::new(),
+            help_controller: HelpController::new(),
             indice_controller: IndiceController::new(),
             fileinput_controller: FileInputController::new(),
             infoaggiuntive_controller: InfoAggiuntiveController::new(),
@@ -65,8 +67,8 @@ impl Controllers {
             CurrentView::Home => {
                 self.home_controller.update(rl, main_state);
             }
-            CurrentView::Second => {
-                self.second_controller.update(rl, main_state);
+            CurrentView::Help => {
+                self.help_controller.update(rl, main_state);
             }
             CurrentView::SelezioneIndice => {
                 self.indice_controller.update(rl, main_state);
@@ -98,57 +100,5 @@ pub(crate) trait Controller {
     fn get_current_index(&self) -> Option<Indice> {
         let state = GLOBAL_STATE.lock().unwrap();
         state.indice_model.get_selected_index()
-    }
-}
-
-// Controller to update and access the state
-
-pub(crate) struct SecondController;
-
-impl Controller for SecondController {
-    type SubModel = SecondModel;
-    fn update(&self, _rl: &mut RaylibHandle, main_state: &mut MainState) {
-        let mut state = GLOBAL_STATE.lock().unwrap();
-        state.second_model.increment_frame_counter();
-        state.second_model.set_name("Updated".to_string());
-        if state.second_model.get_user_continued() {
-            eprintln!("SecondController:  L'utente ha premuto Continua");
-            eprintln!("SecondController:  Let's update current view and go to SelezioneIndice.");
-            main_state.set_current_view(CurrentView::SelezioneIndice)
-        }
-        if main_state.should_reset {
-            eprintln!("SecondController: Resetting");
-            main_state.should_reset = false;
-            state.home_model.reset();
-            state.second_model.reset();
-            state.console_model.reset();
-            main_state.set_current_view(CurrentView::Home);
-            return;
-        }
-    }
-
-    fn get_state(&self) -> Self::SubModel {
-        let state = GLOBAL_STATE.lock().unwrap();
-        state.second_model.clone()
-    }
-}
-
-impl SecondController {
-    pub(crate) fn new() -> Self {
-        Self
-    }
-
-    pub(crate) fn _set_name(&self, name: String) {
-        let mut state = GLOBAL_STATE.lock().unwrap();
-        state.second_model.set_name(name);
-    }
-
-    pub(crate) fn set_value(&self, val: i32) {
-        let mut state = GLOBAL_STATE.lock().unwrap();
-        state.second_model.set_value(val);
-    }
-    pub(crate) fn set_user_continued(&self, val: bool) {
-        let mut state = GLOBAL_STATE.lock().unwrap();
-        state.second_model.set_user_continued(val);
     }
 }